imessage-bots/src/bots/resume-roast/pyproject.toml
[project]
name = "resume-roast"
version = "0.1.0"
description = "A snarky resume roasting chatbot for iMessage via BlueBubbles"
authors = [
    {name = "Shrey Pandya",email = "shrey150@yahoo.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.13,<0.116.0)",
    "uvicorn[standard]>=0.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
